moved admin_annotator class
